rate_limiter: Move rate_limit_rule from test_external and add tests.
